move some test helpers outside of integration_test
